skillsly_service_ms/requirements.txt
django==3.1
djangorestframework==3.11.0
psycopg2-binary==2.9.3